Added test for Plex
I don't want to reboot if Plex is in the top 5 CPU consumers because it means Plex might be serving video.  I don't want to reboot if Plex is serving video.  This will be run in cron so it will (hopefully) reboot the next time there is no connection and Plex isn't running.  Hopefully this won't expose my real IP while Plex is running..?